sreduce trait methods. Assume implementor will figure it out
