+++ src/mesh_lib/millis/mod.rs
@@ -1,4 +1,6 @@
 pub mod traits;
 pub mod types;
 
+pub use traits::PlatformTime;
 pub use types::ms;
+
